# Conflicts:
#	libavdevice/bmd_dec.c

+++ libavcodec/avcodec.h
@@ -2926,7 +2926,7 @@ typedef struct AVCodecContext {
      */
     enum AVPixelFormat sw_pix_fmt;
 
-    uint64_t dropped_frames;
+    int dropped_frames;
 } AVCodecContext;
 
 /**
